+++ src/admin/mod.rs
@@ -1,4 +1,6 @@
-mod admin;
+mod members;
 pub mod tokens;
+mod transfer;
 
-pub use admin::*;
+pub use members::*;
+pub use transfer::*;
